openssl: Modify behavior when running under test
This patch alters the code to fallback to non-FIPS compliant routines
when we detect we are running under a test. This is to allow for us to
use this backend with the Go standard library and also pass all the
tests without invasive modifications to the Go stdlib test functions to
ignore or skip unsupported configurations.