Add v2 module
There are more changes necessary to support Go 1.19 and
many of them are backwards incompatible as there are
function signature changes. This patch introduces a
v2 module which, when tagged, can be used by our Go
fork and allow us to maintain both versions for
pre-Go1.19 and the current Go1.19+ releases.