Use EVP_MD_get_size in all supported OpenSSL versions
As EVP_MD_size is defined as a compatibility macro in OpenSSL 3.0, we
should be using EVP_MD_get_size instead, while it is not present in
OpenSSL 1.1/1.0.  This adds a fallback implementation of
EVP_MD_get_size using EVP_MD_size for those old OpenSSL versions and
switches to using EVP_MD_get_size everywhere.

Signed-off-by: Daiki Ueno <dueno@redhat.com>