Allocate and free on the correct C heaps
In some cases, particularly on Windows [1], the OpenSSL library may have
been linked against a different C runtime library than the rest of the
program. When this is the case, OpenSSL allocates and frees objects on a
different heap than the C heap which cgo and statically-linked C code
allocates on. This is not an issue with objects which are allocated and
freed through typed constructors and destructors exported from OpenSSL,
nor with objects for which we explicitly allocate and free, as those are
consistently freed back to the same heap they were allocated from. But
there are a couple of situations where OpenSSL takes ownership of an
object we allocate, and we take ownership of an object OpenSSL
allocates. Freeing an object onto a different heap than it was allocated
from will corrupt one heap and leak memory from the other. Use the
OpenSSL library functions to allocate and free those objects,
respectively, so as not to corrupt the heaps.

[1]: https://learn.microsoft.com/en-us/cpp/c-runtime-library/potential-errors-passing-crt-objects-across-dll-boundaries?view=msvc-170

Signed-off-by: Cory Snider <csnider@mirantis.com>