rsa: skip setting label for RSA OAEP if empty
The label parameter for RSA OAEP is optional and by default it should
be empty.  Given that it is set only once in setupEVP, we can bypass
the call to EVP_PKEY_CTX_set0_rsa_oaep_label if the label is empty to
spare 1-byte memory allocation and the following CGO call.

Signed-off-by: Daiki Ueno <dueno@redhat.com>