mkcgo: Add missing error symbols to {load,unload}_init_3
When `initForCheckVersion` is called, and if there is an error in
openssl's initialization, the codepath can reach `mkcgo_err_retrieve`
which may invoke `_mkcgo_ERR_get_error_all`.  This will make the
program segfault because the symbol for `ERR_get_error_all` is NULL.

We need to load a couple more symbols from openssl in order to
properly handle the error and display a sane message to the user.

Signed-off-by: Sergio Durigan Junior <sergiodj@chainguard.dev>